// src/apis/models/FireBytes_Unified_Wave_ControlPanelTreeNodes_Dtos_ControlePanelTreeNodeLoadingDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type FireBytes_Unified_Wave_ControlPanelTreeNodes_Dtos_ControlePanelTreeNodeLoadingDto = {
    id?: string;
    nodeId?: string;
    parentNodeId?: string | null;
    nodeName?: string | null;
    icon?: string | null;
    type?: string | null;
    bladeVisualId?: string | null;
    parentType?: string | null;
    parentIdList?: string | null;
    isAction?: boolean;
    isLeaf?: boolean;
    graphicType?: string | null;
    isHaveSlave?: boolean;
    decompressingFilesId?: string | null;
    decompressingFilesName?: string | null;
};